# specs/1-ai-textbook/contracts/openapi.yaml
...
openapi: 3.0.0
info:
  title: Physical AI & Humanoid Robotics Textbook API
  version: 1.0.0
  description: API for managing textbook content, user authentication, RAG chatbot, personalization, and translation features.
servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: http://localhost:8000/v1
    description: Development server
...
paths:
  /auth/signup:
    post:
      summary: User signup
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                background_info:
                  type: object
      responses:
        '201':
          description: User created successfully
  /auth/signin:
    post:
      summary: User signin
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: User signed in successfully
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []